size change for sphere based on distance and scuffed return to controller after disengaging from zipline
as anticipated, it's hard to control when moving

considering scrapping the grapple concept entirely

will sit on it and come back to the project

MVP is basically complete.
It's more of a matter of cutting out the complications if I want to make a presentable MVP.

Wanted to make bioshock infinite skyrails in VR and I have all the mechanics and framework I'd need for that.

Will come back to this project when I have some more design ideas.